Add files via upload
esptool dump of the firmware, going to have an invalid check since it was edited to remove personal information. This clearly shows that the printers UI functions and networking are run on the ESP32 likely meaning that the ARM GD32F103 is running the steppers/extruder. Compared to a RAMPS 1.4 this printer is capable of magnitudes of IOPs.